+++ CHANGELOG.md
@@ -24,6 +24,7 @@ title: Changelog
   * fix tests for PricingConsumptionSimulator
   * rename and update documentation for PricingConsumptionSimulator
   * add `progressive` to PricingConsumptionSimulator
+  * add `secability` to PricingConsumptionSimulator
 
 ## 7.0.0 (2021-09-03)